requirements.txt
torch
torchvision
pandas
numpy
Pillow
tqdm
matplotlib
kagglehub